+++ calmdownman-fanpage/src/components/Header.js
@@ -14,7 +14,7 @@ const Header = () => {
           <StyledImageLink to={path.root} />
           <StyledMenuWrapper $isOpened={isOpened}>
             <StyledLink to={path.teahyeong}>태형</StyledLink>
-            <StyledLink to={path.choonghyun}>충현</StyledLink>
+            <StyledLink to={path.choonghyun}>그의 작품</StyledLink>
             <StyledLink to={path.soyoun}>Scheduler</StyledLink>
             <StyledLink to={path.video}>영상</StyledLink>
           </StyledMenuWrapper>
