--- calmdownman-fanpage/src/pages/choonghyun/Sound.js
@@ -1,30 +1,0 @@
-// import React from 'react';
-// // import playIcon from 경로 설정
-// // import pauseIcon from 경로 설정
-// // import Audio from 경로 설정
-
-// class Music extends React.Component{
-//     constructor(props) {
-//         super(props);
-//         this.state= { 'play': false};
-//         this.s
-//     }
-// }
-
-
-// // ----------------------
-// const myAudio = new Audio();//경로 설정
-// myAudio.src = ""//경로
-
-// myAudio.play();
-// myAudio.pause();
-
-// loop(true);
-
-// // //.audio-player{
-// //     position : fixed;
-// //     top : 20px;
-// //     right: 20px;
-    
-    
-// // }
