+++ calmdownman-fanpage/src/pages/choonghyun/Swip.js
@@ -1,17 +1,8 @@
 import { Swiper, SwiperSlide } from 'swiper/react';
 import 'swiper/css';
 import 'swiper/css/effect-cards';
-import './styles.css';
 import { EffectCards } from 'swiper/modules';
-
-
-const users = [
-  {First: "https://comic.naver.com/webtoon/list?titleId=103759", Last: "https://image-comic.pstatic.net/webtoon/103759/thumbnail/thumbnail_IMAG21_3546920286159910456.jpg"},
-  {First: "https://comic.naver.com/webtoon/list?titleId=602921", Last: "https://image-comic.pstatic.net/webtoon/602921/thumbnail/thumbnail_IMAG21_3847030872581824819.jpg"},
-  {First: "https://comic.naver.com/webtoon/list?titleId=704595", Last: "https://image-comic.pstatic.net/webtoon/704595/thumbnail/thumbnail_IMAG21_3978194032709283938.jpg"},
-  {First: "https://comic.naver.com/webtoon/list?titleId=557527", Last: "https://image-comic.pstatic.net/webtoon/557527/thumbnail/thumbnail_IMAG21_7090408969817175097.jpg"},
-  {First: "https://comic.naver.com/webtoon/list?titleId=687127", Last: "https://image-comic.pstatic.net/webtoon/687127/thumbnail/thumbnail_IMAG21_4063431270715437410.jpg"}
-];
+import users from './Users';
 
 export default function Swip()  {
   return (
